requirements.txt
pyzbar~=0.1.9
Pillow~=11.2.1
qrcode
opencv-python~=4.11.0.86
flask_socketio~=5.3.6
ultralytics~=8.3.153
numpy~=1.23.5
flask~=2.3.3
requests~=2.32.3
tensorflow~=2.19.0
flask==2.3.2
flask-sqlalchemy==3.0.3
flask-socketio==5.3.6
python-socketio~=5.13.0
websocket-client
SpeechRecognition~=3.14.3

psycopg2~=2.9.9
scikit-learn~=1.6.1
PyYAML

torch~=2.7.1+cpu
torchvision~=0.22.1+cpu
efficientnet-pytorch~=0.7.1
matplotlib~=3.7.2
seaborn~=0.13.2
albumentations